feat: post 속성 제거 및 게시판 식별자 추가
